xiaozhi-mcp-bridge/requirements.txt
websockets>=12.0
asyncio-subprocess>=0.1.0
pyyaml>=6.0.1
aiofiles>=23.2.1
paramiko>=3.4.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth>=2.23.0
google-auth-oauthlib>=1.0.0
requests>=2.31.0
notion-client>=2.2.1

